+++ jsonpath/function_extensions/__init__.py
@@ -1,15 +1,15 @@
 # noqa: D104
-from .arguments import validate
-from .count import Count
+from .arguments import validate  # noqa: I001
 from .filter_function import ExpressionType
 from .filter_function import FilterFunction
+from .count import Count
 from .is_instance import IsInstance
 from .keys import keys
-from .length import length
+from .length import Length
 from .match import Match
 from .search import Search
 from .typeof import TypeOf
-from .value import value
+from .value import Value
 
 __all__ = (
     "Count",
@@ -17,10 +17,10 @@ __all__ = (
     "FilterFunction",
     "IsInstance",
     "keys",
-    "length",
+    "Length",
     "Match",
     "Search",
     "TypeOf",
     "validate",
-    "value",
+    "Value",
 )